requirements.txt
selenium==3.141.0
pytest >= 6.1.1
allure-pytest>= 2.8.40
allure-python-commons>=2.8.40
SQLAlchemy == 1.4.11
pymysql == 1.0.2
Faker == 8.2.0
openpyxl == 3.0.7
requests == 2.25.1
python-dotenv == 0.18.0